# build/swagger.json
{
    "components": {
...
        "schemas": {
            "CreateDetailDto": {
                "properties": {
                    "orderId": {
                        "type": "number",
                        "format": "double"
                    },
                    "detailTitle": {
                        "type": "string"
                    },
                    "detailPartNumber": {
                        "type": "string"
                    },
                    "avalabilityInWarehouse": {
                        "type": "boolean"
                    },
                    "detailPrice": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "orderId",
                    "detailTitle",
                    "detailPartNumber",
                    "avalabilityInWarehouse",
                    "detailPrice"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "OrderDto": {
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "double"
                    },
                    "workerId": {
                        "type": "number",
                        "format": "double"
                    },
                    "clientId": {
                        "type": "number",
                        "format": "double"
                    },
                    "orderStatus": {
                        "type": "string"
                    },
                    "orderDescription": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "workerId",
                    "clientId",
                    "orderStatus",
                    "orderDescription"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "OrderStatus": {
                "enum": [
                    "diagnostic",
                    "ordering spare parts",
                    "repair",
                    "ready"
                ],
                "type": "string"
            },
            "CreateOrderDto": {
                "properties": {
                    "workerId": {
                        "type": "number",
                        "format": "double"
                    },
                    "clientId": {
                        "type": "number",
                        "format": "double"
                    },
                    "orderStatus": {
                        "$ref": "#/components/schemas/OrderStatus"
                    },
                    "orderDescription": {
                        "type": "string"
                    }
                },
                "required": [
                    "workerId",
                    "clientId",
                    "orderStatus",
                    "orderDescription"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateOrderPaymentDto": {
                "properties": {
                    "clientId": {
                        "type": "number",
                        "format": "double"
                    },
                    "orderId": {
                        "type": "number",
                        "format": "double"
                    },
                    "detailPriceSum": {
                        "type": "number",
                        "format": "double"
                    },
                    "workPrice": {
                        "type": "number",
                        "format": "double"
                    },
                    "paymentConformation": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "clientId",
                    "orderId",
                    "detailPriceSum",
                    "workPrice",
                    "paymentConformation"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "car_repair_station",
        "version": "1.0.0",
        "description": "Back-END car repair station",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Stawr28@gmail.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
        "/details": {
            "get": {
                "operationId": "GetAllDetails",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateDetailDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "detail"
                ],
                "security": [],
                "parameters": []
            },
            "post": {
                "operationId": "CreateDetail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateDetailDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "detail"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDetailDto"
                            }
                        }
                    }
                }
            }
        },
        "/details/{detailId}": {
            "get": {
                "operationId": "GetDetail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateDetailDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "detail"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "detailId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateDetail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateDetailDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "detail"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "detailId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDetailDto"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteDetail",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "detail"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "detailId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "operationId": "GetAllOrders",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDto"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "CreateOrder",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderDto"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{orderId}": {
            "get": {
                "operationId": "GetOrder",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker",
                            "client"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "orderId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateOrder",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin",
                            "worker"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "orderId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderDto"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteOrder",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "barearAuth": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "orderId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        },
        "/orderPayment": {
            "get": {
                "operationId": "GetAllPayment",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderPaymentDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "orderPayment"
                ],
                "security": [],
                "parameters": []
            },
            "post": {
                "operationId": "CreateOrderPayment",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderPaymentDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "orderPayment"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderPaymentDto"
                            }
                        }
                    }
                }
            }
        },
        "/orderPayment/{orderPaymentId}": {
            "get": {
                "operationId": "GetOrderPayment",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderPaymentDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "orderPayment"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "orderPaymentId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateOrderPayment",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateOrderPaymentDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "orderPayment"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "orderPaymentId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderPaymentDto"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}